backend/requirements.txt
fastapi==0.116.1
motor==3.7.1
pydantic==2.11.7
pydantic-settings==2.10.1
PyJWT==2.10.1
pydantic_core==2.33.2
pymongo==4.14.1
python-dotenv==1.1.1
uvicorn==0.35.0
email-validator==2.3.0
asyncpg==0.30.0
reportlab==4.4.3
